Update Ver 0.0.12

+++ start.sh
@@ -5,12 +5,12 @@ export PATH
 #=================================================
 #	System Required: CentOS 6+/Debian 6+/Ubuntu 14.04+
 #	Description: Test the vps and install the software
-#	Version: 0.0.11
+#	Version: 0.0.12
 #	Author: VPSBASH
 #	Email: VPSBASH@Gmail.com
 #	#Scripts copy from the big guys
 #=================================================
-sh_ver="0.0.11"
+sh_ver="0.0.12"
 Info="${Green_font_prefix}[信息]${Font_color_suffix}"
 Error="${Red_font_prefix}[错误]${Font_color_suffix}"
 Tip="${Green_font_prefix}[注意]${Font_color_suffix}"
@@ -502,7 +502,8 @@ LotServer(){
  ${Green_font_prefix}6.${Font_color_suffix} 重启 LotServer
  ${Green_font_prefix}7.${Font_color_suffix} 查看 LotServer 状态
  
- 注意： 目前脚本不支持Ubuntu18.04和Debian9" && echo
+ 注意:	目前脚本不支持Debian9 
+		删除内核时候请选择NO" && echo
 	stty erase '^H' && read -p "(默认: 取消):" lotserver_num
 	[[ -z "${lotserver_num}" ]] && echo "已取消..." && BBRLos
 	if [[ ${lotserver_num} == "1" ]]; then
@@ -536,27 +537,25 @@ Replace_kernel(){
 	check_root
     check_sys
 if [[ ${release} == "ubuntu" ]]; then
-	[ -n "`cat /etc/issue | grep "Ubuntu 18.04"`" ] && echo "Ubuntu 18.04不支持安装锐速" && exit 0
+	[ -n "`cat /etc/issue | grep "Ubuntu 18.04"`" ] && Ubuntu18_kernel
 	[ -n "`cat /etc/issue | grep "Ubuntu 16.04"`" ] && echo "Ubuntu 16.04" && KER_VER="4.4.0-47-generic"
 	[ -n "`cat /etc/issue | grep "Ubuntu 14.04"`" ] && echo "Ubuntu 14.04" && KER_VER="3.16.0-43-generic"
 	sudo apt-get update
 	apt-get install -y linux-image-extra-$KER_VER linux-image-$KER_VER
-	del_kernel
-	sudo update-grub
-	reboot
-	exit 0
+	if [ -n "`dpkg -l | grep "linux-image-extra-$KER_VER"`" ]; then
+		del_kernel
+		sudo update-grub
+		reboot
+		exit 0
+	else
+		echo -e "${Error} 内核安装失败，请检查你的网络和软件源"
+	fi
+
 elif [[ ${release} == "debian" ]]; then
 	[ -n "`cat /etc/issue | grep "Linux 9"`" ] && echo "Debian 9不支持安装锐速" && exit 0
-	[ -n "`cat /etc/issue | grep "Linux 8"`" ] && echo "Linux 8" && KER_VER="3.16.0-4-amd64-dbg" && D_VER="jessie"
+	[ -n "`cat /etc/issue | grep "Linux 8"`" ] && echo "Linux 8" && Debian8_kernel
 	[ -n "`cat /etc/issue | grep "Linux 7"`" ] && Debian7_kernel
-	cp /etc/apt/sources.list /etc/apt/sources.list_bak
-	echo -e "\ndeb http://ftp.debian.org/debian/ $D_VER-backports main" >> /etc/apt/sources.list
-	apt-get update
-	apt-get -t $D_VER-backports install linux-image-$KER_VER-$arch -y
-	del_kernel
-	update-grub
-	reboot
-	exit 0
+
 elif [[ ${release} == "centos" ]]; then
 	[ -n "`cat /etc/redhat-release |grep "CentOS Linux release 7"`" ] &&
 	yum -y install linux-firmware &&
@@ -565,10 +564,43 @@ elif [[ ${release} == "centos" ]]; then
 	rpm -ivh http://soft.91yun.org/ISO/Linux/CentOS/kernel/kernel-firmware-2.6.32-504.3.3.el6.noarch.rpm &&
 	rpm -ivh http://soft.91yun.org/ISO/Linux/CentOS/kernel/kernel-2.6.32-504.3.3.el6.x86_64.rpm --force
 	reboot
+
 else 
 	echo -e "${Error} 您的操作系统未在支持列表内" && exit 0
 	fi
 exit 0
+}
+Ubuntu18_kernel(){
+[ -n "`cat /etc/issue | grep "Ubuntu 18.04"`" ] && echo "Linux 8" && KER_VER="4.4.0-47"
+[ -n "`uname -m | grep "x86_64"`" ] && echo "amd64" && arch="amd64"
+[ -n "`uname -m | grep "i686"`" ] && echo "i686" && arch="i386"
+	wget http://mirrors.kernel.org/ubuntu/pool/main/l/linux/linux-headers-4.4.0-47_4.4.0-47.68_all.deb
+	wget http://mirrors.kernel.org/ubuntu/pool/main/l/linux/linux-headers-4.4.0-47-generic_4.4.0-47.68_$arch.deb
+	wget http://mirrors.kernel.org/ubuntu/pool/main/l/linux/linux-image-4.4.0-47-generic_4.4.0-47.68_$arch.deb
+	sudo dpkg -i linux-headers-4.4.0*.deb linux-image-4.4.0*.deb
+if [ -n "`dpkg -l | grep "linux-image-$KER_VER"`" ]; then
+	del_kernel
+	update-grub
+	reboot
+	exit 0
+else
+	echo -e "${Error} 内核安装失败，请检查你的网络和软件源"
+	fi
+}
+Debian8_kernel(){
+[ -n "`cat /etc/issue | grep "Linux 8"`" ] && echo "Linux 8" && KER_VER="3.16.0-4-amd64-dbg" && D_VER="jessie"
+	cp /etc/apt/sources.list /etc/apt/sources.list_bak
+	echo -e "\ndeb http://ftp.debian.org/debian/ $D_VER-backports main" >> /etc/apt/sources.list
+	apt-get update
+	apt-get -t $D_VER-backports install linux-image-$KER_VER -y
+	if [ -n "`dpkg -l | grep "linux-image-$KER_VER"`" ]; then
+	del_kernel
+	update-grub
+	reboot
+	exit 0
+else
+	echo -e "${Error} 内核安装失败，请检查你的网络和软件源"
+	fi
 }
 Debian7_kernel(){
 	[ -n "`cat /etc/issue | grep "Linux 7"`" ] && echo "Linux 7" && KER_VER="3.2.0-4" && D_VER="wheezy"
@@ -578,10 +610,14 @@ Debian7_kernel(){
 	echo -e "\ndeb http://ftp.debian.org/debian/ $D_VER-backports main" >> /etc/apt/sources.list
 	apt-get update
 	apt-get -t $D_VER-backports install linux-image-$KER_VER-$arch -y
+	if [ -n "`dpkg -l | grep "linux-image-$KER_VER-$arch"`" ]; then
 	del_kernel
 	update-grub
 	reboot
 	exit 0
+else
+	ehco -e "${Error} 内核安装失败，请检查你的网络和软件源"
+	fi
 }
 del_kernel(){
 	deb_total=`dpkg -l | grep linux-image | awk '{print $2}' | grep -v "${KER_VER}" | wc -l`
@@ -637,6 +673,7 @@ Uninstall_LotServer(){
 		echo && echo "LotServer 卸载完成 !" && echo
 	fi
 }
+
 BBR_2(){
     check_sys
 if [[ ${release} == "centos" ]]; then
